moved contentful hocs from test-site into organisms
